+++ BackEnd/src/main/java/com/manhattan/busyness_predictor/repository/SharedRepository.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Repository;
 import com.manhattan.busyness_predictor.model.Shared;
 
 @Repository
-public interface SharedRepository extends JpaRepository<Shared, Long> {
+public interface SharedRepository extends JpaRepository<Shared, Integer> {
 
-    List<Shared> findByReceiverIdOrderBySharedAtDesc(Long receiverId);
+    List<Shared> findByReceiverIdOrderBySharedAtDesc(Integer receiverId);
 
-    List<Shared> findBySenderIdOrderBySharedAtDesc(Long senderId);
+    List<Shared> findBySenderIdOrderBySharedAtDesc(Integer senderId);
 }
